Update to Documentation
Added notations (fill colors) to the spreadsheet for the M65C02B tab to document a new branch mode that incorporates OSX to add capability to support C/Pascal boolean evaluation and tests. Also added notes detailing the reasoning behind the branch instructions and changes to the SP-relative addressing mode and the operand order of SP-relative subtraction instructions.